{"schema":"decodex/commit/1","summary":"Split MCP control modules","authority":"XY-1115"}
